update_plugins_json/requirements.txt
requests
git-python
pygithub
